feat: added support for encrypting events persisted in sqlite3 db (#118)
* feat: added support for encrypting events persisted in sqlite3 db

* chore: updated pubspec.lock files across all packages w.r.t to latest dependencies

* chore: improved sample app

* chore: updated sample app to use the encryption feature

* chore: updated ios plugin to use the latest version of iOS SDK

* fix: fixed issues for encryption configuration on android platform

* refactor: moved DBEncryption to a separate class

* fix: fixed dbencryption status to be sent even if it was disabled to the native sdks

* chore: updated podfile.lock of example app

---------
