+++ packages/plugins/rudder_plugin_ios/ios/Classes/RudderSdkFlutterPlugin.m
@@ -184,6 +184,17 @@ BOOL isRegistrarDetached = NO;
   if ([dataResidencyServer isEqualToString:@"EU"]) {
       [configBuilder withDataResidencyServer:EU];
   }
+
+  if([configDict objectForKey:@"dbEncryption"]) {
+    NSDictionary* dbEncryptionDict = [configDict objectForKey:@"dbEncryption"];
+    BOOL enabled = [[dbEncryptionDict objectForKey:@"enabled"] boolValue];
+    NSString* encryptionKey = [dbEncryptionDict objectForKey:@"key"];
+
+    if (encryptionKey != nil && [encryptionKey length] != 0) {
+        [configBuilder withDBEncryption:[[RSDBEncryption alloc] initWithKey:encryptionKey enable:enabled]];
+    }
+  }
+
   if (integrationList != nil) {
     for (id<RSIntegrationFactory> integration in integrationList) {
       [configBuilder withFactory:integration];
